feat(GIFT-12948): unit test improvements
