+++ src/modules/ods/ods.service.test.ts
@@ -1,6 +1,4 @@
-import { InternalServerErrorException, NotFoundException } from '@nestjs/common';
-import { BUSINESS_CENTRE, CUSTOMERS, DEALS } from '@ukef/constants';
-import { mapBusinessCentres } from '@ukef/helpers';
+import { CUSTOMERS } from '@ukef/constants';
 import { PinoLogger } from 'nestjs-pino';
 import { DataSource, QueryRunner } from 'typeorm';
 
@@ -26,340 +24,21 @@ describe('OdsService', () => {
     service = new OdsService(mockDataSource, mockLogger);
   });
 
-  describe('findCustomer', () => {
-    const mockCustomer = { urn: CUSTOMERS.EXAMPLES.PARTYURN, name: 'Test Customer' };
-
-    const mockStoredProcedureOutput = `{
-      "query_request_id": "Test ID",
-      "message": "SUCCESS",
-      "status": "SUCCESS",
-      "total_result_count": 1,
-      "results": [
-        {
-          "customer_party_unique_reference_number": "${mockCustomer.urn}",
-          "customer_name": "${mockCustomer.name}",
-          "customer_companies_house_number": "12345678",
-          "customer_addresses": [
-            {
-              "customer_address_type": "Registered",
-              "customer_address_street": "Test Street",
-              "customer_address_postcode": "AA1 1BB",
-              "customer_address_country": "United Kingdom",
-              "customer_address_city": "Test City"
-            }
-          ]
-        }
-      ]
-    }`;
-
-    beforeEach(() => {
-      jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-    });
-
-    it('should call service.callOdsStoredProcedure', async () => {
-      await service.findCustomer(mockCustomer.urn);
-
-      const expectedStoredProcedureInput: OdsStoredProcedureInput = service.createOdsStoredProcedureInput({
-        entityToQuery: ODS_ENTITIES.CUSTOMER,
-        queryPageSize: 1,
-        queryParameters: { customer_party_unique_reference_number: mockCustomer.urn },
-      });
-
-      expect(service.callOdsStoredProcedure).toHaveBeenCalledTimes(1);
-      expect(service.callOdsStoredProcedure).toHaveBeenCalledWith(expectedStoredProcedureInput);
-    });
-
-    it('should return a customer', async () => {
-      const result = await service.findCustomer(mockCustomer.urn);
-
-      expect(result).toEqual(mockCustomer);
-    });
-
-    describe('when the response from ODS does not have status as SUCCESS', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "NOT SUCCESS" }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('Error finding a customer');
-
-        const promise = service.findCustomer(mockCustomer.urn);
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when the response from ODS has total_result_count as 0', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "SUCCESS", "total_result_count": 0 }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('No customer found');
-
-        const promise = service.findCustomer(mockCustomer.urn);
-
-        await expect(promise).rejects.toBeInstanceOf(NotFoundException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when the method goes into the catch handler', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "SUCCESS" }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('Error finding a customer');
-
-        const promise = service.findCustomer(mockCustomer.urn);
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when callOdsStoredProcedure throws an error', () => {
-      it('should throw an error', async () => {
-        jest.spyOn(service, 'callOdsStoredProcedure').mockRejectedValue('Mock ODS error');
-
-        const expected = new Error('Error finding a customer');
-
-        const promise = service.findCustomer(mockCustomer.urn);
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-  });
-
-  describe('findDeal', () => {
-    const mockStoredProcedureOutput = `{
-      "query_request_id": "Test ID",
-      "message": "SUCCESS",
-      "status": "SUCCESS",
-      "total_result_count": 1,
-      "results": [
-        {
-          "deal_code": "${DEALS.EXAMPLES.ID}",
-          "deal_name": "${DEALS.EXAMPLES.NAME}",
-          "deal_type_description": "${DEALS.EXAMPLES.DESCRIPTION}"
-        }
-      ]
-    }`;
-
-    beforeEach(() => {
-      jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-    });
-
-    it('should call service.callOdsStoredProcedure', async () => {
-      const expectedStoredProcedureInput: OdsStoredProcedureInput = service.createOdsStoredProcedureInput({
-        entityToQuery: ODS_ENTITIES.DEAL,
-        queryPageSize: 1,
-        queryParameters: { deal_code: DEALS.EXAMPLES.ID },
-      });
-
-      await service.findDeal(DEALS.EXAMPLES.ID);
-
-      expect(service.callOdsStoredProcedure).toHaveBeenCalledTimes(1);
-      expect(service.callOdsStoredProcedure).toHaveBeenCalledWith(expectedStoredProcedureInput);
-    });
-
-    it('should return a deal', async () => {
-      const result = await service.findDeal(DEALS.EXAMPLES.ID);
-
-      const expected = {
-        dealId: DEALS.EXAMPLES.ID,
-        name: DEALS.EXAMPLES.NAME,
-        description: DEALS.EXAMPLES.DESCRIPTION,
-      };
-
-      expect(result).toEqual(expected);
-    });
-
-    describe('when the response from ODS does not have status as SUCCESS', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "NOT SUCCESS" }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('Error finding a deal');
-
-        const promise = service.findDeal(DEALS.EXAMPLES.ID);
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when the response from ODS has total_result_count as 0', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "SUCCESS", "total_result_count": 0 }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('No deal found');
-
-        const promise = service.findDeal(DEALS.EXAMPLES.ID);
-
-        await expect(promise).rejects.toBeInstanceOf(NotFoundException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when the method goes into the catch handler', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "SUCCESS" }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('Error finding a deal');
-
-        const promise = service.findDeal(DEALS.EXAMPLES.ID);
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when callOdsStoredProcedure throws an error', () => {
-      it('should throw an error', async () => {
-        jest.spyOn(service, 'callOdsStoredProcedure').mockRejectedValue('Mock ODS error');
-
-        const expected = new Error('Error finding a deal');
-
-        const promise = service.findDeal(DEALS.EXAMPLES.ID);
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-  });
-
-  describe('getBusinessCentres', () => {
-    const mockStoredProcedureOutput = `{
-      "message": "SUCCESS",
-      "status": "SUCCESS",
-      "total_result_count": 2,
-      "results": [
-        {
-          "business_centre_code": "${BUSINESS_CENTRE.EXAMPLES.CODE}",
-          "business_centre_name": "${BUSINESS_CENTRE.EXAMPLES.NAME}"
-        },
-        {
-          "business_centre_code": "${BUSINESS_CENTRE.EXAMPLES.CODE}",
-          "business_centre_name": "${BUSINESS_CENTRE.EXAMPLES.NAME}"
-        }
-      ]
-    }`;
-
-    beforeEach(() => {
-      jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-    });
-
-    it('should call service.callOdsStoredProcedure', async () => {
-      const expectedStoredProcedureInput: OdsStoredProcedureInput = service.createOdsStoredProcedureInput({
-        entityToQuery: ODS_ENTITIES.BUSINESS_CENTRE,
-      });
-
-      await service.getBusinessCentres();
-
-      expect(service.callOdsStoredProcedure).toHaveBeenCalledTimes(1);
-      expect(service.callOdsStoredProcedure).toHaveBeenCalledWith(expectedStoredProcedureInput);
-    });
-
-    it('should return mapped business centres', async () => {
-      const result = await service.getBusinessCentres();
-
-      const expected = mapBusinessCentres(JSON.parse(mockStoredProcedureOutput).results);
-
-      expect(result).toEqual(expected);
-    });
-
-    describe('when the response from ODS does not have status as SUCCESS', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "NOT SUCCESS" }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('Error getting business centres');
-
-        const promise = service.getBusinessCentres();
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when the response from ODS has total_result_count as 0', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "SUCCESS", "total_result_count": 0 }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('Error getting business centres');
-
-        const promise = service.getBusinessCentres();
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when the method goes into the catch handler', () => {
-      it('should throw an error', async () => {
-        const mockStoredProcedureOutput = `{ "status": "SUCCESS" }`;
-
-        jest.spyOn(service, 'callOdsStoredProcedure').mockResolvedValue(mockStoredProcedureOutput);
-
-        const expected = new Error('Error getting business centres');
-
-        const promise = service.getBusinessCentres();
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-
-    describe('when callOdsStoredProcedure throws an error', () => {
-      it('should throw an error', async () => {
-        jest.spyOn(service, 'callOdsStoredProcedure').mockRejectedValue('Mock ODS error');
-
-        const expected = new Error('Error getting business centres');
-
-        const promise = service.getBusinessCentres();
-
-        await expect(promise).rejects.toBeInstanceOf(InternalServerErrorException);
-
-        await expect(promise).rejects.toThrow(expected);
-      });
-    });
-  });
-
   describe('createOdsStoredProcedureInput', () => {
     it('should map the inputs to the stored procedure input format', () => {
+      // Arrange
       const exampleCustomerQueryParameters = {
         customer_party_unique_reference_number: CUSTOMERS.EXAMPLES.PARTYURN,
       };
 
+      // Act
       const result = service.createOdsStoredProcedureInput({
         entityToQuery: ODS_ENTITIES.CUSTOMER,
         queryPageSize: 100,
         queryParameters: exampleCustomerQueryParameters,
       });
 
+      // Assert
       const expected: OdsStoredProcedureInput = {
         query_method: 'get',
         query_object: ODS_ENTITIES.CUSTOMER,
@@ -374,6 +53,7 @@ describe('OdsService', () => {
 
   describe('callOdsStoredProcedure', () => {
     it('should call the stored procedure with the query runner', async () => {
+      // Arrange
       const mockInput: OdsStoredProcedureInput = service.createOdsStoredProcedureInput({
         entityToQuery: ODS_ENTITIES.CUSTOMER,
         queryPageSize: 100,
@@ -386,8 +66,10 @@ describe('OdsService', () => {
 
       mockQueryRunner.query.mockResolvedValue(mockResult);
 
+      // Act
       const result = await service.callOdsStoredProcedure(mockInput);
 
+      // Assert
       expect(mockDataSource.createQueryRunner).toHaveBeenCalledTimes(1);
 
       expect(mockQueryRunner.query).toHaveBeenCalledWith(
@@ -403,6 +85,7 @@ describe('OdsService', () => {
     });
 
     it('throws an error if calling the stored procedure fails', async () => {
+      // Arrange
       const mockInput: OdsStoredProcedureInput = service.createOdsStoredProcedureInput({
         entityToQuery: ODS_ENTITIES.CUSTOMER,
         queryParameters: { customer_party_unique_reference_number: CUSTOMERS.EXAMPLES.PARTYURN },
@@ -410,9 +93,10 @@ describe('OdsService', () => {
 
       mockQueryRunner.query.mockRejectedValue(new Error('Test Error'));
 
-      const resultPromise = service.callOdsStoredProcedure(mockInput);
+      // Act & Assert
+      const promise = service.callOdsStoredProcedure(mockInput);
 
-      await expect(resultPromise).rejects.toThrow('Test Error');
+      await expect(promise).rejects.toThrow('Test Error');
 
       expect(mockDataSource.createQueryRunner).toHaveBeenCalledTimes(1);
       expect(mockQueryRunner.release).toHaveBeenCalledTimes(1);
